Isolate common types for AstarteFlow
